requirements.txt
pyTelegramBotAPI==4.9.0
mysql-connector-python==8.0.32
Pillow==9.4.0
googletrans==4.0.0rc1